+++ src/Kiota.Builder/KiotaBuilder.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -1553,7 +1553,7 @@ public class KiotaBuilder
             var parentClassNamespace = GetShortestNamespace(currentNamespace, parentSchema);
             inheritsFrom = (CodeClass)AddModelDeclarationIfDoesntExist(currentNode, parentSchema, parentSchema.GetSchemaName().CleanupSymbolName(), parentClassNamespace);
         }
-        var newClass = currentNamespace.AddClass(new CodeClass
+        var newClassStub = new CodeClass
         {
             Name = declarationName,
             Kind = CodeClassKind.Model,
@@ -1563,9 +1563,16 @@ public class KiotaBuilder
                 DocumentationLink = schema.ExternalDocs?.Url,
                 Description = schema.Description.CleanupDescription(),
             },
-        }).First();
+        };
         if (inheritsFrom != null)
-            newClass.StartBlock.Inherits = new CodeType { TypeDefinition = inheritsFrom, Name = inheritsFrom.Name };
+            newClassStub.StartBlock.Inherits = new CodeType { TypeDefinition = inheritsFrom, Name = inheritsFrom.Name };
+        
+        // Add the class to the namespace after the serialization members
+        // as other threads looking for the existence of the class may find the class but the additional data/backing store properties may not be fully populated causing duplication
+        var includeAdditionalDataProperties = config.IncludeAdditionalData && schema.AdditionalPropertiesAllowed;
+        AddSerializationMembers(newClassStub, includeAdditionalDataProperties, config.UsesBackingStore);
+
+        var newClass = currentNamespace.AddClass(newClassStub).First();
         CreatePropertiesForModelClass(currentNode, schema, currentNamespace, newClass); // order matters since we might be recursively generating ancestors for discriminator mappings and duplicating additional data/backing store properties
 
         var mappings = GetDiscriminatorMappings(currentNode, schema, currentNamespace, newClass)
@@ -1708,11 +1715,6 @@ public class KiotaBuilder
     }
     private void CreatePropertiesForModelClass(OpenApiUrlTreeNode currentNode, OpenApiSchema schema, CodeNamespace ns, CodeClass model)
     {
-
-        var includeAdditionalDataProperties = config.IncludeAdditionalData &&
-            (schema?.AdditionalPropertiesAllowed ?? false);
-
-        AddSerializationMembers(model, includeAdditionalDataProperties, config.UsesBackingStore);
         if (schema?.Properties?.Any() ?? false)
         {
             model.AddProperty(schema